Moving routes to mock_data.py and insights.py
